Added README.md to symbit crate
