+++ crates/sym/Cargo.toml
@@ -13,5 +13,9 @@ rust-version.workspace = true
 
 [dependencies]
 thiserror.workspace = true
-pcode-ops = { path = "../pcode-ops/" }
-aiger-circuit = "1"
+aiger-circuit = { version = "1", optional = true }
+
+[features]
+
+# Enables integration with aiger-circuit for generating AIGER formatted circuit
+aiger = ["dep:aiger-circuit"]
